Move log inserting logic to new test helper class, LogInserter.
